+++ lib/src/channels/channel.dart
@@ -7,7 +7,7 @@ abstract class Channel {
   const Channel(this.options);
 
   /// Listen for an event on the channel instance.
-  Channel listen(String event, Function callback);
+  void listen(String event, Function callback);
 
   /// subscribe channel
   void subscribe();
@@ -16,25 +16,25 @@ abstract class Channel {
   void unsubscribe();
 
   /// Listen for a whisper event on the channel instance.
-  Channel listenForWhisper(String event, Function callback) =>
+  void listenForWhisper(String event, Function callback) =>
       listen('.client-$event', callback);
 
   /// Listen for an event on the channel instance.
-  Channel notification(Function callback) => listen(
+  void notification(Function callback) => listen(
         '.Illuminate\\Notifications\\Events\\BroadcastNotificationCreated',
         callback,
       );
 
   /// Stop listening to an event on the channel instance.
-  Channel stopListening(String event, [Function? callback]);
+  void stopListening(String event, [Function? callback]);
 
   /// Stop listening for a whisper event on the channel instance.
-  Channel stopListeningForWhisper(String event, [Function? callback]) =>
+  void stopListeningForWhisper(String event, [Function? callback]) =>
       stopListening('.client-$event', callback);
 
   /// Register a callback to be called anytime a subscription succeeds.
-  Channel subscribed(Function callback);
+  void subscribed(Function callback);
 
   /// Register a callback to be called anytime an error occurs.
-  Channel error(Function callback);
+  void error(Function callback);
 }