+++ lib/src/channels/private_channel.dart
@@ -7,8 +7,8 @@ abstract class PrivateChannel extends Channel {
   const PrivateChannel(super.options);
 
   /// Register a callback to be called anytime the member list changes.
-  PrivateChannel whisper(String eventName, dynamic data);
+  void whisper(String eventName, dynamic data);
 
   /// listen to on subscribe success event
-  Channel onSubscribedSuccess(Function callback);
+  void onSubscribedSuccess(Function callback);
 }