+++ lib/src/channels/pusher/pusher_presence_channel.dart
@@ -1,6 +1,6 @@
 import 'package:pusher_client_socket/pusher_client_socket.dart' as PUSHER;
 
-import 'presence_channel.dart';
+import '../presence_channel.dart';
 import 'pusher_channel.dart';
 
 ///
@@ -16,20 +16,17 @@ class PusherPresenceChannel extends PusherChannel<PUSHER.PresenceChannel>
 
   /// Register a callback to be called anytime the member list changes.
   @override
-  PusherPresenceChannel here(Function callback) =>
-      this..on('pusher:subscription_succeeded', callback);
+  void here(Function callback) => on('pusher:subscription_succeeded', callback);
 
   /// Listen for someone joining the channel.
   @override
-  PusherPresenceChannel joining(Function callback) =>
-      this..on('pusher:member_added', callback);
+  void joining(Function callback) => on('pusher:member_added', callback);
 
   /// Listen for someone leaving the channel.
   @override
-  PusherPresenceChannel leaving(Function callback) =>
-      this..on('pusher:member_removed', callback);
+  void leaving(Function callback) => on('pusher:member_removed', callback);
 
   /// Trigger client event on the channel.
-  PusherPresenceChannel whisper(String eventName, dynamic data) =>
-      this..subscription.trigger('client-$eventName', data);
+  void whisper(String eventName, dynamic data) =>
+      subscription.trigger('client-$eventName', data);
 }