+++ lib/src/connector/connector.dart
@@ -1,6 +1,13 @@
-import '../channel/channel.dart';
-import '../channel/presence_channel.dart';
-import '../channel/private_channel.dart';
+import 'dart:convert';
+
+import 'package:pinenacl/api.dart';
+import 'package:pinenacl/x25519.dart' show SecretBox;
+import 'package:socket_io_client/socket_io_client.dart' as IO;
+
+import '../channels/channel.dart';
+import '../channels/presence_channel.dart';
+import '../channels/private_channel.dart';
+import '../channels/private_encrypted_channel.dart';
 
 abstract class Connector<ClientType, ChannelType> {
   /// Connector options.
@@ -25,9 +32,9 @@ abstract class Connector<ClientType, ChannelType> {
   PresenceChannel presenceChannel(String channel);
 
   /// listen to channel event
-  Channel listen(String channel, String event, Function callback);
+  void listen(String channel, String event, Function callback);
 
-  Channel? encryptedPrivateChannel(String channel) => null;
+  PrivateEncryptedChannel privateEncryptedChannel(String channel);
 
   /// Leave the given channel, as well as its private and presence variants.
   void leave(String channel);
@@ -88,4 +95,56 @@ class ConnectorOptions<T> {
     this.authHeaders,
     this.nameSpace,
   });
+}
+
+class SocketIoConnectorOptions extends ConnectorOptions<IO.Socket> {
+  /// The channel decryption handler.
+  final Map<String, dynamic> Function(
+    String sharedSecret,
+    Map data,
+  )? channelDecryption;
+
+  const SocketIoConnectorOptions({
+    required super.client,
+    super.authHeaders,
+    super.nameSpace,
+    this.channelDecryption,
+  });
+
+  ByteList _decodeCipherText(String cipherText) {
+    Uint8List uint8list = base64Decode(cipherText);
+    ByteData byteData = ByteData.sublistView(uint8list);
+    List<int> data = List<int>.generate(
+        byteData.lengthInBytes, (index) => byteData.getUint8(index));
+    return ByteList(data);
+  }
+
+  Map<String, dynamic> decryptChannelData(
+    String sharedSecret,
+    Map data,
+  ) =>
+      (channelDecryption ?? defaultChannelDecryptionHandler)(
+        sharedSecret,
+        data,
+      );
+
+  Map<String, dynamic> defaultChannelDecryptionHandler(
+    String sharedSecret,
+    Map data,
+  ) {
+    if (!data.containsKey("ciphertext") || !data.containsKey("nonce")) {
+      throw Exception(
+        "Unexpected format for encrypted event, expected object with `ciphertext` and `nonce` fields, got: $data",
+      );
+    }
+
+    final ByteList cipherText = _decodeCipherText(data["ciphertext"]);
+
+    final Uint8List nonce = base64Decode(data["nonce"]);
+
+    final SecretBox secretBox = SecretBox(base64Decode(sharedSecret));
+    final Uint8List decryptedData = secretBox.decrypt(cipherText, nonce: nonce);
+
+    return jsonDecode(utf8.decode(decryptedData)) as Map<String, dynamic>;
+  }
 }